requirements.txt
# Core dependencies for Nana's Resume Builder
beautifulsoup4==4.12.2
requests==2.31.0
Pillow==10.1.0
pytesseract==0.3.10
PyPDF2==3.0.1
spacy==3.7.2
click==8.1.7
python-docx==1.1.0
reportlab==4.0.7
weasyprint==60.1
pyyaml==6.0.1
python-dotenv==1.0.0

# Optional: For AI integration (if using OpenAI)
openai==1.3.0

# Development tools
pytest==7.4.3
black==23.11.0
flake8==6.1.0